Put the converter from string to MathNode in a separate class
